ADEN-12740 Move getDataSettingsFromMetatag to shared
