+++ src/platforms/news-and-ratings/shared/index.ts
@@ -4,3 +4,4 @@ export * from './setup/context/lazy-loaded-slots-context.setup';
 export * from './setup/context/targeting/news-and-ratings-targeting.setup';
 export * from './setup/dynamic-slots/news-and-ratings-dynamic-slots.setup';
 export * from './setup/wad/news-and-ratings-wad.setup';
+export * from './utils/get-data-settings-from-metatag';
